+++ tests/test_pytest.py
@@ -17,8 +17,7 @@
 from __future__ import annotations
 
 from extra_platforms import (
-    LINUX,
-    current_os,
+    is_linux,
     is_macos,
     is_windows,
 )
@@ -34,38 +33,38 @@ from extra_platforms.pytest import (
 
 @skip_linux
 def test_skip_linux():
-    assert current_os() not in LINUX
+    assert not is_linux()
     assert is_macos() or is_windows()
 
 
 @skip_macos
 def test_skip_macos():
     assert not is_macos()
-    assert current_os() in LINUX or is_windows()
+    assert is_linux() or is_windows()
 
 
 @skip_windows
 def test_skip_windows():
     assert not is_windows()
-    assert current_os() in LINUX or is_macos()
+    assert is_linux() or is_macos()
 
 
 @unless_linux
 def test_unless_linux():
-    assert current_os() in LINUX
+    assert is_linux()
     assert not is_macos()
     assert not is_windows()
 
 
 @unless_macos
 def test_unless_macos():
-    assert current_os() not in LINUX
+    assert not is_linux()
     assert is_macos()
     assert not is_windows()
 
 
 @unless_windows
 def test_unless_windows():
-    assert current_os() not in LINUX
+    assert not is_linux()
     assert not is_macos()
     assert is_windows()